requirements.txt
fastapi
uvicorn[standard]
crewai
crewai[google-genai]
python-dotenv
pydantic
